fix(gui): close the localization audit on passes 9+10 — #678 Phase 4
An audit of the branch's ten new or reworded keys found two
blockers and three majors in the ENGLISH, before a word was
translated. Each is fixed at the call site, English being the
source `scripts/extract-keys` reads.

The blockers:

- `profiles.broken.malformed` is the first string in the corpus
  to put "JSON" inside a sentence a non-Latin locale must render,
  and `GLOSSARY` had no `json` — so `english_residue` would have
  refused the ja/ko/ru/zh merges. The branch was green only
  because no locale had translated it yet. Fixed where
  localization.md says a term that stays English belongs, in the
  glossary, not as an exemption.
- The inherited-row VoiceOver hint said "base layer". The
  inheritance axis is base-vs-profile WITHIN the selected layer,
  and "base layer" is a different real thing on the same screen
  (`shortcuts.base_layer_protected`). Spoken-only, so no sighted
  reviewer would ever have caught it. Now "base shortcuts",
  matching the caption users read.

The majors, all one shape — a string contradicting another string
on its own screen:

- `app_bar.override.axlabel` was a coined twin of the visible
  column header `space_override.override_column`. Deleted; the
  checkbox takes the header's own key, so the name VoiceOver
  speaks is the word on screen in all ten languages.
- `palettes.empty_hint` called the built-ins "what colors
  KiwiDesk", which `palettes.caption` two rows up rules out ("a
  one-time paint, not a live link"), and offered to save one of
  them, which `palettes.reserved` forbids.
- `shortcuts.apps.empty` said a key "brings it forward"; the
  authoritative `shortcuts.app_behavior.help` says it brings the
  window to the CURRENT SPACE. Crossing spaces is the whole
  distinguishing behaviour.

And the twin the audit only flagged: `settings.permission_paused`
`.action` held the same English as `onboarding.grant.open_settings`
— two keys, ten chances to drift, and `onboarding.grant.body`
QUOTES that label in its numbered steps. Collapsed into one
`common.open_system_settings` rather than pinned by a guard: a
rename carries every existing translation over, and one key
cannot disagree with itself.

Minors: the three `profiles.broken.*` sentences no longer repeat
the header directly above them; `borders.preview.axlabel` becomes
`border.preview.ax` (the only plural stem in a 34-key singular
family) and names its subject, a bare "Preview:" being a preview
of nothing in the VoiceOver rotor; contractions in
`monitors.none_detected`.

`KeybindingAppGroup.swift` split at the §2.1 ceiling — the row
renderer and its two record paths move to `+Row`, on the seam
`+AddRow` and `+Behavior` already use. `SettingsAnchorPrimitiveTests`
reds on that move, exactly as a fail-shut allow-list should: the
row's collection `.id()` now lives in the new file.
