+++ Sources/KiwiDesk/Settings/Components/Keybindings/KeybindingAppGroup.swift
@@ -10,10 +10,12 @@ import SwiftUI
 struct ApplicationsGroup: View {
     @ObservedObject var model: SettingsModel
     @Binding var bindings: [KeyBinding]
+    /// Read by the row in `+Row` (dimming) and by the recorders
+    /// there (the live-apply seam), so neither is `private`.
     @Environment(\.keybindingOverrideBase)
-    private var overrideBase
+    var overrideBase
     @Environment(\.keybindingLayerName)
-    private var layerName
+    var layerName
     /// The app chosen in the add-row but not yet committed — the
     /// row only enters the list once an app is picked, so no
     /// app-less placeholder can exist (matches App Rules). Read by
@@ -34,17 +36,24 @@ struct ApplicationsGroup: View {
         ) {
             if orderedAppIDs.isEmpty {
                 // What holds instead of the empty list (#678
-                // Phase 4 pass 9, turn 18). "No app shortcuts" is
-                // a fact the reader can see; that an app opens or
-                // comes forward on the SAME key is the thing the
-                // emptiness leaves them wondering about, since
-                // that behaviour is what a binding here changes.
+                // Phase 4 pass 9, turn 18): what a binding here
+                // changes is what the emptiness leaves the reader
+                // wondering about. Worded FROM
+                // `shortcuts.app_behavior.help`, authoritative
+                // for the Open or Focus default a new row gets —
+                // crossing SPACES is the distinguishing
+                // behaviour, "brings it forward" is plain
+                // activation, and the two would take different
+                // verbs in ten languages (l10n audit,
+                // 2026-08-11).
                 Text(
                     L(
                         "shortcuts.apps.empty",
                         "No app has a key of its own yet. Add one "
-                            + "and it opens that app, or brings it "
-                            + "forward when it is already running."
+                            + "and its key brings that app's "
+                            + "window to the space you're on, or "
+                            + "launches the app if it isn't "
+                            + "running."
                     )
                 )
                 .font(.caption)
@@ -117,121 +126,7 @@ struct ApplicationsGroup: View {
         )
     }
 
-    private func row(
-        _ binding: Binding<KeyBinding>
-    ) -> some View {
-        HStack {
-            // Tighter than the row's ambient spacing so "app + how
-            // to open it" reads as one unit (ui-designer, #334).
-            HStack(spacing: 6) {
-                appMenu(binding)
-                behaviorMenu(binding)
-            }
-            Spacer()
-            KeyRecorderField(
-                combo: binding.wrappedValue.combo,
-                conflict: ConflictText.tooltip(
-                    for: binding.wrappedValue,
-                    in: bindings,
-                    config: model.config
-                ),
-                preflight: { combo in
-                    RecorderPreflight.rejection(
-                        combo: combo,
-                        excluding: {
-                            [id = binding.wrappedValue.id] in
-                            $0.id == id
-                        },
-                        bindings: $bindings,
-                        // Id-based (#68 review M2): Steal
-                        // mutates the array before this runs.
-                        commit: {
-                            _ = record(
-                                $0,
-                                id: binding.wrappedValue.id
-                            )
-                        }
-                    )
-                },
-                onRecord: { record($0, into: binding) },
-                onClear: {
-                    let id = binding.wrappedValue.id
-                    binding.wrappedValue.combo = ""
-                    // Live target: unregister now (#123).
-                    _ = model.liveApplyRecorded(
-                        layerName: layerName,
-                        bindingID: id,
-                        combo: nil
-                    )
-                }
-            )
-            Button {
-                remove(binding.wrappedValue.id)
-            } label: {
-                Image(systemName: "trash")
-            }
-            .buttonStyle(.borderless)
-            .iconButtonAffordance(
-                L(
-                    "shortcuts.remove_binding",
-                    "Remove shortcut"
-                )
-            )
-        }
-        .keybindingRowStyle(
-            inherited: binding.wrappedValue.isInherited(
-                from: overrideBase
-            )
-        )
-        .id(binding.wrappedValue.id.uuidString)
-    }
-
-    private func record(
-        _ combo: String,
-        into binding: Binding<KeyBinding>
-    ) -> LiveApplyFeedback? {
-        binding.wrappedValue.combo = combo
-        let id = binding.wrappedValue.id
-        if let index = bindings.firstIndex(
-            where: { $0.id == id }
-        ) {
-            model.noteRecordedCombo(
-                bindings[index],
-                in: bindings
-            )
-        }
-        return model.liveApplyRecorded(
-            layerName: layerName,
-            bindingID: id,
-            combo: combo
-        )
-    }
-
-    /// Looks the row up by id at write time — safe after any
-    /// structural mutation of the bindings array.
-    @discardableResult
-    private func record(
-        _ combo: String,
-        id: UUID
-    ) -> LiveApplyFeedback? {
-        guard
-            let index = bindings.firstIndex(where: {
-                $0.id == id
-            })
-        else { return nil }
-        bindings[index].combo = combo
-        model.noteRecordedCombo(
-            bindings[index],
-            in: bindings
-        )
-        return model.liveApplyRecorded(
-            layerName: layerName,
-            bindingID: id,
-            combo: combo
-        )
-    }
-
-    private func appMenu(
+    func appMenu(
         _ binding: Binding<KeyBinding>
     ) -> some View {
         AppPickerButton(
@@ -332,7 +227,9 @@ struct ApplicationsGroup: View {
     /// Revert — `liveApplyRecorded` rebuilds the running table
     /// from its own session copy, which never saw the removal.
     /// No-op off the live target, where nothing is registered.
-    private func remove(_ id: UUID) {
+    /// Called by the row's trash button in `+Row`, so not
+    /// `private`.
+    func remove(_ id: UUID) {
         bindings.removeAll { $0.id == id }
         _ = model.liveApplyRecorded(
             layerName: layerName,
