+++ Sources/KiwiDeskCore/Resources/Locales/en.json
@@ -58,7 +58,6 @@
   "app_bar.liquid_glass": "Liquid Glass",
   "app_bar.liquid_glass.help": "Lays a translucent glass material over the boxes or the plate. The Fill color tints it, though the tint reads subtle on current macOS.",
   "app_bar.no_layout.help": "No layout shows an App Bar — turn one on below.",
-  "app_bar.override.axlabel": "Override",
   "app_bar.override.off.help": "Inheriting the global value",
   "app_bar.override.on.help": "Overriding the global value",
   "app_bar.preview.app_code": "Code",
@@ -163,6 +162,7 @@
   "border.glow_size": "Glow size",
   "border.glow_size.auto": "Auto glow size",
   "border.glow_size.disabled": "Turn on Glow effect to adjust its size.",
+  "border.preview.ax": "Border preview: a focused window with its ring, beside an unfocused one.",
   "border.shared.caption": "The focus ring, the drag ghost and the drop zone are strokes KiwiDesk draws around a window.",
   "border.shared.differ.help": "The three strokes are set differently right now; choosing here sets all three.",
   "border.shared.title": "Shared by all borders",
@@ -170,7 +170,6 @@
   "border.unfocused_color.a11y": "Unfocused window border color",
   "border.unfocused_enabled": "Show border on unfocused windows",
   "border.width": "Width",
-  "borders.preview.axlabel": "Preview: a focused window with its ring, beside an unfocused one.",
   "brand.menu_bar_icon.a11y": "KiwiDesk",
   "color_field.automatic": "Automatic",
   "color_field.hex.a11y": "%1$@ hex value",
@@ -200,6 +199,7 @@
   "colors.unfocused_off.help": "Unfocused windows get no border, so this color isn't drawn. Turn it on in %1$@.",
   "common.off": "off",
   "common.on": "on",
+  "common.open_system_settings": "Open System Settings",
   "config_issues.cancel": "Cancel",
   "config_issues.delete": "Delete…",
   "config_issues.delete_confirm.body": "This removes the profile file. It can't be undone.",
@@ -633,7 +633,7 @@
   "monitors.area.caption": "Drag a space onto the display it belongs to.",
   "monitors.clear_pin.label": "Back to automatic placement",
   "monitors.deselect": "Hide the spaces on this display",
-  "monitors.none_detected": "No monitors detected, so there is nothing to place spaces on. Windows still tile, in a single space.",
+  "monitors.none_detected": "No monitors detected, so there's nothing to place spaces on. Windows still tile, all in a single space.",
   "monitors.not_connected": "Monitors not connected",
   "monitors.not_connected.caption": "This profile's monitors aren't attached right now, so its space placement can't be edited here. Connect its monitor setup to arrange spaces — the other sections still edit this profile.",
   "monitors.orphan_pin.row_axlabel": "%1$@, waiting for monitor %2$@",
@@ -668,7 +668,6 @@
   "onboarding.discovery.body": "Press %1$@ anytime to see an overview of all your shortcuts.",
   "onboarding.grant.body": "1. Click “Open System Settings” below.\n2. Find KiwiDesk in the list and turn it on.\n3. Come back here — we detect it automatically.",
   "onboarding.grant.granted": "Permission granted!",
-  "onboarding.grant.open_settings": "Open System Settings",
   "onboarding.grant.title": "Enable Accessibility",
   "onboarding.grant.waiting": "Waiting for permission…",
   "onboarding.ready.body": "KiwiDesk is already arranging your windows with sensible defaults. Open Settings and apply a preset for your setup — then fine-tune anything from there. You can always get back here from the menu bar — or press ⌃⌥K for a shortcuts cheat-sheet.",
@@ -688,7 +687,7 @@
   "palettes.caption": "Apply a bundled or saved set of colors to the App Bar, Space Bar, borders, and drag visuals — a one-time paint, not a live link.",
   "palettes.default_name": "My Palette",
   "palettes.delete": "Delete",
-  "palettes.empty_hint": "The built-ins above are what colors KiwiDesk. Save one and it joins them here.",
+  "palettes.empty_hint": "The palettes above paint KiwiDesk in one click. Save colors of your own and they appear here.",
   "palettes.export": "Export…",
   "palettes.import": "Import…",
   "palettes.mine": "My palettes",
@@ -763,10 +762,10 @@
   "profiles.area.caption": "A profile is your whole setup, remembered per display arrangement.",
   "profiles.badge.active": "active",
   "profiles.badge.default": "default",
-  "profiles.broken.malformed": "Couldn't load — the file isn't valid JSON. Opening it will show where it broke.",
+  "profiles.broken.malformed": "Not valid JSON — opening it will show where it broke.",
   "profiles.broken.title": "Couldn't load",
-  "profiles.broken.unexpected_shape": "Couldn't load — saved by another version, or a hand edit changed one of its fields.",
-  "profiles.broken.unreadable": "Couldn't be read — the file may have been moved or deleted.",
+  "profiles.broken.unexpected_shape": "Saved by another version, or a hand edit changed one of its fields.",
+  "profiles.broken.unreadable": "The file may have been moved or deleted.",
   "profiles.copy_failed": "Copying failed: %1$@",
   "profiles.current_setup_note": "Your current setup is saved into %1$@. To keep it separately, use \"Save a Copy As…\" in the bar below.",
   "profiles.delete.help": "Delete profile",
@@ -850,7 +849,6 @@
   "settings.globals_save_failed": "Saving settings failed: %1$@",
   "settings.layout.drift": "Live: %1$@ — profile has %2$@ (not saved)",
   "settings.permission_paused": "Window management is paused — KiwiDesk needs Accessibility permission.",
-  "settings.permission_paused.action": "Open System Settings",
   "settings.readout.auto": "Automatic",
   "shortcuts.add": "Add",
   "shortcuts.add_application": "Add application",
@@ -863,7 +861,7 @@
   "shortcuts.app_behavior.open_new": "Open New",
   "shortcuts.app_behavior.open_new.badge": "Opens a new instance",
   "shortcuts.app_behavior.open_or_focus": "Open or Focus",
-  "shortcuts.apps.empty": "No app has a key of its own yet. Add one and it opens that app, or brings it forward when it is already running.",
+  "shortcuts.apps.empty": "No app has a key of its own yet. Add one and its key brings that app's window to the space you're on, or launches the app if it isn't running.",
   "shortcuts.base_layer_protected": "Base layers can't be removed here",
   "shortcuts.choose_app": "Choose app…",
   "shortcuts.delete_layer": "Delete layer",
@@ -897,7 +895,7 @@
   "shortcuts.remove_binding": "Remove shortcut",
   "shortcuts.rename": "Rename",
   "shortcuts.rename_ellipsis": "Rename…",
-  "shortcuts.row.inherited.axhint": "Inherited from the base layer. Record a key here to override it.",
+  "shortcuts.row.inherited.axhint": "Inherited from the base shortcuts. Record a key here to override it.",
   "shortcuts.section.focus": "Focus",
   "shortcuts.section.general": "General",
   "shortcuts.section.inactive": "Inactive shortcuts",
